Consolidate Extensions sidebar stories
